% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoint_funs.R
\name{api_endpoints}
\alias{api_endpoints}
\alias{publishUI}
\alias{view_api}
\alias{do_encoded}
\alias{ui}
\alias{publishPlot}
\title{TITLE TBD}
\usage{
publishUI(ui = NULL, script = NULL, style = NULL, path = NULL)

view_api(path)

do_encoded(fun, args)

ui(html = NULL)

publishPlot(hxp, width, height, res)
}
\arguments{
\item{ui}{TBD}

\item{script}{TBD}

\item{style}{TBD}

\item{path}{TBD}

\item{fun}{TBD}

\item{args}{TBD}

\item{html}{TBD}

\item{hxp}{TBD}

\item{width}{TBD}

\item{height}{TBD}

\item{res}{TBD}
}
\description{
TITLE TBD
}
\section{Functions}{
\itemize{
\item \code{publishUI}: TBD

\item \code{view_api}: TBD

\item \code{do_encoded}: TBD

\item \code{ui}: TBD

\item \code{publishPlot}: TBD
}}

