% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_utils.R
\name{auth_utils}
\alias{auth_utils}
\alias{ocpu_cipher}
\alias{ocpu_decipher}
\alias{ocpu_unserialize}
\alias{ocpu_serialize}
\alias{valid_client_apikey}
\title{OCPU Auth Utils}
\usage{
ocpu_cipher(msg)

ocpu_decipher(ciphx, privx)

ocpu_unserialize(ocpu_json, local = FALSE)

ocpu_serialize(ll, pubx = NULL)

valid_client_apikey(keyx, noncex, privx, hash)
}
\arguments{
\item{msg}{TBD}

\item{ciphx}{TBD}

\item{privx}{TBD}

\item{ocpu_json}{TBD}

\item{local}{TBD}

\item{ll}{TBD}

\item{pubx}{TBD}

\item{keyx}{TBD}

\item{noncex}{TBD}

\item{hash}{TBD}
}
\description{
OCPU Auth Utils
}
\section{Functions}{
\itemize{
\item \code{ocpu_cipher}: simple encryption using server's pub key

\item \code{ocpu_decipher}: simple decryption using server's priv key

\item \code{ocpu_unserialize}: TBD

\item \code{ocpu_serialize}: TBD

\item \code{valid_client_apikey}: TBD
}}

