% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpu_print.R
\name{ocpu_print}
\alias{ocpu_print}
\alias{printstamp}
\alias{printmsg}
\alias{print_ocpu_call}
\title{OCPU PRINT UTILITY FUNCTIONS}
\usage{
printstamp(msg = "")

printmsg(msg, sym = "+", slim = TRUE, content = NULL,
  addtime = TRUE)

print_ocpu_call(decoded)
}
\arguments{
\item{msg}{TBD}

\item{sym}{TBD}

\item{slim}{TBD}

\item{content}{TBD}

\item{addtime}{TBD}

\item{decoded}{TBD}
}
\description{
OCPU PRINT UTILITY FUNCTIONS
}
\section{Functions}{
\itemize{
\item \code{printstamp}: TBD

\item \code{printmsg}: TBD

\item \code{print_ocpu_call}: TBD
}}

