% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_api.R
\name{plot_api}
\alias{plot_api}
\alias{plot2hex}
\alias{hex2plot}
\alias{plotpost}
\title{Title}
\usage{
plot2hex(p)

hex2plot(hxp)

plotpost(p, h = 500, w = 500, res = 98)
}
\arguments{
\item{p}{TBD}

\item{hxp}{TBD}

\item{h}{TBD}

\item{w}{TBD}

\item{res}{TBD}
}
\description{
Title
}
\section{Functions}{
\itemize{
\item \code{plot2hex}: TBD

\item \code{hex2plot}: TBD

\item \code{plotpost}: TBD
}}

