% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{api_utils}
\alias{api_utils}
\alias{ocpu_api}
\alias{ocpu_body}
\alias{ocpu_icecube_url}
\alias{ocpu_radix_url}
\alias{wrap_args}
\alias{ocpu_parsed_post}
\title{OCPU API Utils}
\usage{
ocpu_api(pkg, fun, verb, host, ...)

ocpu_body(...)

ocpu_icecube_url(fun)

ocpu_radix_url(apikey, fun)

wrap_args(...)

ocpu_parsed_post(url, args = NULL, config = NULL)
}
\arguments{
\item{pkg}{TBD}

\item{fun}{TBD}

\item{verb}{TBD}

\item{host}{TBD}

\item{...}{TBD}

\item{apikey}{TBD}

\item{url}{TBD}

\item{args}{TBD}

\item{config}{TBD}
}
\description{
OCPU API Utils
}
\section{Functions}{
\itemize{
\item \code{ocpu_api}: TBD

\item \code{ocpu_body}: TBD

\item \code{ocpu_icecube_url}: TBD

\item \code{ocpu_radix_url}: TBD

\item \code{wrap_args}: TBD

\item \code{ocpu_parsed_post}: TBD
}}

