% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpu_calls.R
\name{ocpu_calls}
\alias{ocpu_calls}
\alias{encode_call}
\alias{decode_call}
\title{TBD}
\usage{
encode_call(fun, args)

decode_call(encoded)
}
\arguments{
\item{fun}{TBD}

\item{args}{TBD}

\item{encoded}{TBD}
}
\description{
TBD
}
\section{Functions}{
\itemize{
\item \code{encode_call}: TBD

\item \code{decode_call}: TBD
}}

