% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpu_print.R
\name{ocpu_print}
\alias{ocpu_print}
\alias{printstamp}
\alias{printmsg}
\title{OCPU PRINT UTILITY FUNCTIONS}
\usage{
printstamp(msg = "")

printmsg(msg, sym = "+", slim = TRUE, content = NULL,
  addtime = TRUE)
}
\arguments{
\item{msg}{TBD}

\item{sym}{TBD}

\item{slim}{TBD}

\item{content}{TBD}

\item{addtime}{TBD}
}
\description{
OCPU PRINT UTILITY FUNCTIONS
}
\section{Functions}{
\itemize{
\item \code{printstamp}: TBD

\item \code{printmsg}: TBD
}}

