% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_api.R
\name{plot_api}
\alias{plot_api}
\alias{plot2hex}
\alias{hex2plot}
\title{Title}
\usage{
plot2hex(p)

hex2plot(hxp)
}
\arguments{
\item{p}{TBD}

\item{hxp}{TBD}
}
\description{
Title
}
\section{Functions}{
\itemize{
\item \code{plot2hex}: TBD

\item \code{hex2plot}: TBD
}}

